Split V3App.svelte into focused components
Closes #203. V3App was ~2.3k LoC mixing orchestration, ghost rendering +
d3 physics, focus FSM, drag gestures, pending-family bookkeeping, and
modal wiring. Now ~487 LoC of pure composition.

Extracted modules / components:

- pendingState.ts — pending* types + helpers (id-gen, descriptions,
  composeFamilyMembers).
- v3DomGeometry.ts — clientToSvgPoint, nodeCenter, findNodeUnder.
- v3MutationActions.ts — class owning withPendingAdd / promote /
  attach / createPersonInFamily / onGhostClick / runRemove*. Reactive
  state stays in V3App as $state; the class mutates it through
  getter/setter refs so plain-TS modules don't need runes.
- v3ModalActions.ts — addStemma / renameStemma / cloneStemma /
  removeStemma / addPerson / confirmRemove* flows.
- v3Session.ts — Google OAuth + e2e auto-login lifecycle.
- V3FocusController.svelte — desktop hover-to-focus effect.
- V3GhostLayer.svelte — ghost branch rendering + d3 ghost sim
  (freeze/unfreeze neighbours, snap-back on blur), click dispatch.
- V3DragGestures.svelte — v2 drag-matrix gestures (1-6 from
  frontend/CLAUDE.md), HTML5 DnD, drop classification, drag tip,
  pan/space/escape keys.
- V3EditModeOverlay.svelte — edit-mode watermark + diagonal pattern.
- V3Chrome.svelte — top/bottom chrome layers and bulk-add tray.
- V3PendingVisuals.svelte — pending-add / pending-remove circle
  classes effect.

No behaviour change — v3 e2e smoke + v2 e2e suite + 195 frontend unit
tests pass; svelte-check clean.
